Fix a bunch of warnings

+++ src/test/java/org/assertj/core/api/Assertions_assertThat_inBinary_Test.java
@@ -75,14 +75,14 @@ public class Assertions_assertThat_inBinary_Test {
   @Test
   public void should_assert_long_in_binary() {
     thrown.expectMessage("expected:<...000_00000000_0000001[1]> but was:<...000_00000000_0000001[0]>");
-    assertThat((long) 2).inBinary().isEqualTo((long) 3);
+    assertThat((long) 2).inBinary().isEqualTo(3);
   }
 
   @Test
   public void should_assert_negative_long_in_binary() {
     thrown.expectMessage("expected:<0b[00000000_00000000_00000000_00000000_00000000_00000000_00000000_00000011]> " +
         "but was:<0b[11111111_11111111_11111111_11111111_11111111_11111111_11111111_11111110]>");
-    assertThat((long) -2).inBinary().isEqualTo((long) 3);
+    assertThat((long) -2).inBinary().isEqualTo(3);
   }
 
   @Test
